Tidy: TensorLayout::from, Engine::layout, RateAdaptor cleanup, smaller fixes
- TensorLayout::from(InferenceConfig, state_pairs) translates anira's
  config into a layout. Session's constructor body drops from ~70 lines
  of inline loops + cryptic locals to a one-line member-init.
- Engine: four layout forwarders + state_pair_count() replaced by a
  single Engine::layout() returning the underlying TensorLayout&.
- Session: is_state_input/is_state_output removed (live on TensorLayout
  already).
- AniraTilde::init_external: shared label() lambda + msg_in_base /
  msg_out_base named indices in place of four parallel loops.
- Session::prepare: compute_decoder_latency(layout, buffer_size) helper
  returning std::optional<DecoderLatency>; if-let style replaces the
  open-coded loop + flag.
- Mixer::push_dry_sample / mix_wet_sample: int channel → size_t channel.
  Drops the static_cast<int>(channel) at the Engine→Mixer boundary.
- RateAdaptor: per-tensor state co-located in PerTensor struct
  (kind + upsample_pos + downsample_hold + downsample_pop); pre_dispatch
  returns AniraView bundling the four pointer arrays anira needs (drops
  the input_views/output_views/sample-count accessor zoo);
  next_inference_offset() helper names the upsample boundary math.

Tests: still 33/33.

+++ include/anira_tilde/mixing/Mixer.h
@@ -30,8 +30,8 @@ public:
                                size_t frames, size_t channel);
 
     // Per-sample API (kept for unit tests / fine-grained use).
-    void  push_dry_sample(float dry_sample, int channel);
-    float mix_wet_sample (float wet_sample, int channel);
+    void  push_dry_sample(float dry_sample, size_t channel);
+    float mix_wet_sample (float wet_sample, size_t channel);
 
     void   set_mix(float new_mix);
     size_t get_latency() const { return m_latency_samples; }
